+++ package.json
@@ -1,10 +1,8 @@
 {
   "name": "trianglify",
-  "version": "4.0.0-next.2",
+  "version": "4.0.0-next.3",
   "description": "Trianglify is a javascript library for generating colorful triangle meshes that can be used as SVG images and CSS backgrounds.",
   "main": "dist/trianglify.js",
-  "browser": "dist/trianglify.umd.js",
-  "module": "dist/trianglify.module.js",
   "scripts": {
     "lint": "standard",
     "test": "jest",
@@ -39,6 +37,9 @@
       "**/vendor/**"
     ]
   },
+  "jest": {
+    "setupFiles": ["jest-canvas-mock"]
+  },
   "author": "Quinn Rohlf <qr@qrohlf.com>",
   "license": "GPL-3.0",
   "bugs": {
@@ -54,6 +55,7 @@
     "@rollup/plugin-node-resolve": "^7.1.3",
     "babel-eslint": "^10.1.0",
     "jest": "^25.5.4",
+    "jest-canvas-mock": "^2.2.0",
     "rollup": "^2.7.2",
     "rollup-plugin-bundle-size": "^1.0.3",
     "rollup-plugin-terser": "^5.3.0",
